- The JSON to embedding converter and schema validator have been moved to the 'utils' file. - The previous files have been renamed to 'test_fromJsonToEmb' and 'test_schemaValidator'.
