+++ ChatSQL/code/test_fromJsonToEmb.py
@@ -1,32 +1,5 @@
 import os
-import json
-
-def generateEmbeddingUpsert(jsonFileName):
-    with open(jsonFileName, 'r') as file:
-        data = json.load(file)
-
-    commands = []
-    index_counter = 0 # Contatore globale per il numero incrementale
-
-    for table in data["tables"]:
-        table_name = table["name"]
-        table_description = table["table-description"]
-
-        for column in table["columns"]:
-            field_name = column["name"]
-            type = column["type"]
-            references = column["references"]
-            description = column["description"]
-
-            # Create the emb.upsert command
-            dictionary = {"table": table_name, "table-description": table_description,"field": field_name, "type": type, "references": references, "description": description}
-            command = (index_counter, dictionary)
-
-            commands.append(command)
-
-            index_counter += 1
-
-    return commands
+from utils import generateEmbeddingUpsert
 
 if __name__ == "__main__":
     dir_path = os.path.dirname(os.path.realpath(__file__))
