+++ ChatSQL/code/utils.py
@@ -2,6 +2,8 @@ import os
 import sys
 import csv
 import time
+import json
+from jsonschema import validate, ValidationError
 
 dirPath = os.path.dirname(os.path.realpath(__file__))
 sys.path.append(os.path.abspath(os.path.join(dirPath, '..')))
@@ -44,3 +46,38 @@ def loading_animation(n):
 
 if __name__ == "__main__":
     loading_animation(2)
+
+
+def generateEmbeddingUpsert(jsonFileName):
+    with open(jsonFileName, 'r') as file:
+        data = json.load(file)
+
+    commands = []
+    index_counter = 0 # Contatore globale per il numero incrementale
+
+    for table in data["tables"]:
+        table_name = table["name"]
+        table_description = table["table-description"]
+
+        for column in table["columns"]:
+            field_name = column["name"]
+            type = column["type"]
+            references = column["references"]
+            description = column["description"]
+
+            # Create the emb.upsert command
+            dictionary = {"table": table_name, "table-description": table_description,"field": field_name, "type": type, "references": references, "description": description}
+            command = (index_counter, dictionary)
+
+            commands.append(command)
+
+            index_counter += 1
+
+    return commands
+
+def jsonValidator(json_data, json_schema):
+    try:
+        validate(instance=json_data, schema=json_schema)
+        return True, None
+    except ValidationError as e:
+        return False, str(e)
